Improvement at design and change directory

+++ dashboard/admin/admin.php
@@ -5,7 +5,7 @@
 
     //mengecek apakah ada session admin yang aktif, jika tidak maka diarahkan ke login.php
     if(!isset($_SESSION['data']['role']) || $_SESSION['data']['role'] !== 'Admin'){
-        header("location: ../login/"); // arahkan ke login.php
+        header("location: ../../auth/login/"); // arahkan ke login.php
         exit();
     }
 ?>
@@ -16,7 +16,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="../../assets/css/global.css">
-    <link rel="stylesheet" href="../../assets/css/dashboard.css">
+    <link rel="stylesheet" href="../../assets/css/admin.css">
     <link rel="shortcut icon" href="../../assets/img/logo-wp-circle.png">
     <title>Dashboard - Bengkel Wahyu Putra</title>
 </head>
@@ -59,25 +59,20 @@
             <?php while ($result = mysqli_fetch_assoc($sql)) { ?>
             <div class="card">
                 <h2><?= $result['nama_pelanggan']; ?></h2>
-                <h3>@<?= $result['username']; ?></h3>
                 <p><i class="fas fa-phone"></i> <?= $result['no_telp']; ?></p>
-                <p><i class="fas fa-person"></i> <?= $result['jenis_kelamin']; ?></p>
-                <p><i class="fas fa-calendar-days"></i> <?= $result['tgl_lahir']; ?></p>
+                <p><i class="fas fa-envelope"></i> <?= $result['email']; ?></p>
                 <em class="role <?= $result['role']; ?>"><?= $result['role']; ?></em>
                 <em class="type <?= $result['jenis_akun']; ?>"><?= $result['jenis_akun']; ?></em>
 
                 <div class="action-btn">
                     <a href="kelola.php?ubah=<?= $result['id_pelanggan']; ?>" class="btn edit"><i class="fas fa-pen-to-square"></i> Edit</a>
-                    <a href="../proses.php?hapus=<?= $result['id_pelanggan']; ?>" class="btn hapus"><i class="fas fa-trash"></i> Hapus</a>
+                    <a href="../../auth/proses.php?hapus=<?= $result['id_pelanggan']; ?>" class="btn hapus" onclick="return confirm('Apakah Anda yakin untuk menghapus data tersebut?')"><i class="fas fa-trash"></i> Hapus</a>
                 </div>
             </div>
             <?php } ?>
         </div>
     </section>
 
-    <!-- FOOTER -->
-    <?php include "../../includes/footer.php"; ?>
-    <!-- FOOTER END -->
     <script src="https://kit.fontawesome.com/ed13b1bb03.js" crossorigin="anonymous"></script>
     <script src="../../assets/js/main.js"></script>
 </body>
